requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-ollama>=0.0.1
chromadb>=0.4.22
streamlit>=1.30.0
pypdf>=3.17.0
python-dotenv>=1.0.0
sentence-transformers>=2.3.0
unstructured>=0.12.0
